+++ apps/api/.env.example
@@ -293,6 +293,8 @@ PARKING_USER_ID=
 
 # mailchimp setup
 MAILCHIMP_MAIN_LIST_ID=
+MAILCHIMP_PROBELESEN_AUDIENCE_ID=
+MAILCHIMP_ONBOARDING_AUDIENCE_ID=
 MAILCHIMP_INTEREST_NEWSLETTER_DAILY=
 MAILCHIMP_INTEREST_NEWSLETTER_WEEKLY=
 MAILCHIMP_INTEREST_NEWSLETTER_FEUILLLETON=
